pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_m6_project"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Levente Murgas", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "dvc>=3.66.0",
    "dvc-gdrive>=3.0.1",
    "dvc-gs>=3.0.2",
    "evidently>=0.7.20",
    "fastapi==0.115.6",
    "gdown>=5.2.1",
    "google-api-python-client>=2.187.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "kagglehub>=0.4.0",
    "locust>=2.43.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pandas>=2.3.3",
    "prometheus-client>=0.24.1",
    "ptflops>=0.7.5",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "torch==2.9.1",
    "torchao>=0.15.0",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.6",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "mypy>=1.19.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F", "W", "D"]
ignore = ["D100", "D211", "D212", "D401", "D103", "D104", "D101", "D102", "D107"]

[tool.coverage.run]
omit = ["tests/*",
    "*/tmp/*",
    "*/_remote_module_non_scriptable.py",
    "/private/*",
]

[project.scripts]
train = "mlops_m6_project.train:train"
trainl = "mlops_m6_project.train_lightning:train"
evaluate = "mlops_m6_project.evaluate:evaluate"
iris = "mlops_m6_project.iris_classifier:app"
